+++ openfisca_core/periods/__init__.py
@@ -29,6 +29,6 @@ And for classes::
 
 from ._config import INSTANT_PATTERN
 from ._units import DateUnit, DAY, ETERNITY, MONTH, YEAR
-from .helpers import build_period, key_period_size, parse_period
+from .helpers import build_period, parse_period
 from .instant_ import Instant
 from .period_ import Period